+++ src/wdio/services/TerraService.js
@@ -1,4 +1,5 @@
 import chai from 'chai';
+import axeCommand from './TerraCommands/axe-command';
 import chaiMethods from './TerraCommands/chai-methods';
 import accessiblity from './TerraCommands/accessiblity';
 import visualRegression from './TerraCommands/visual-regression';
@@ -50,6 +51,7 @@ export default class TerraService {
 
   // eslint-disable-next-line class-methods-use-this
   before() {
+    global.browser.addCommand('axe', axeCommand);
     chai.config.showDiff = false;
     global.expect = chai.expect;
     global.should = chai.should();