+++ src/wdio/services/TerraCommands/accessiblity.js
@@ -1,11 +1,15 @@
 /**
 * A mocha-chai convenience test case to assert assesiblity.
-* @property {Object} options - The Axe options. Available options are viewports,
-* rules, runOnly, and contex. See https://www.axe-core.org/docs/.
+* @param {Object} options - The Axe options. Available options are viewports,
+* rules, and contex. See https://www.axe-core.org/docs/.
 */
 const beAccessible = (options) => {
   global.it('is accessible', () => {
-    global.expect(global.browser.axe(options)).to.be.accessible();
+    const axeOptions = {
+      context: global.browser.options.terra.selector,
+      ...options,
+    };
+    global.expect(global.browser.axe(axeOptions)).to.be.accessible();
   });
 };